+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Fixed
+
+- Expand config entry button covering the end of the text input
+
 ## 0.7.4 (2024-08-09)
 
 ### Fixed